# api.yaml
openapi: "3.0.2"
info:
  title: Challenge Backend
  version: "1.0"
servers:
  - url: http://{domain}:{port}
    variables:
      domain:
        default: localhost
      port:
        enum:
          - '3000'
          - '8080'
        default: '3000'
  - url: http://localhost:3000/
    description: Local test instance
  - url: http://10.239.68.77:3000
    description: Playground Server

security:
  - bearerAuth: []

paths:

  /scores/{challenge}:
    get:
      summary: Get all scores
      operationId: getScores
      tags:
        - scores
      parameters:
        - name: challenge
          in: path
          description: Challenge name
          required: true
          schema:
            type: string
          example: "23_3_1"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Score'


  /submit/{challenge}:
    post:
      summary: Submit a run
      operationId: submitRun
      tags:
        - run
      parameters:
        - name: challenge
          in: path
          description: Challenge name
          required: true
          schema:
            type: string
          example: "23_3_1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Submit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Score'

components:
  schemas:
    Submit: 
      type: object
      properties:
        player:
          type: string
        name:
          type: string
        language:
          type: string
        code:
          type: string
        test:
          type: boolean

    
    Score:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        time:
          type: integer
          format: int64
        language:
          type: string
        binary:
          type: string

  